The calculation method for intermediary points coordinates is finally working, I checked the output values

+++ tests/moving/position.h
@@ -5,7 +5,7 @@
 
 extern volatile int x_pos;
 extern volatile int y_pos;
-extern volatile int orientation;
+extern volatile double orientation;
 
 void update_position(void);
